BackEnd/requirements.txt
flask==2.3.2
flask-cors==3.0.10
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.5
joblib==1.3.1
snowflake-connector-python==3.16.0
uvicorn==0.35.0